+++ assets/js/web-components/prpl-badge-progress-bar.js
@@ -44,10 +44,10 @@ customElements.define(
 							badge-id="${ badgeId }"
 							style="
 								display:flex;
-								width: 5rem;
-								height: 5rem;
+								width: 7.5rem;
+								height: auto;
 								position: absolute;
-								left: calc(${ progress }% - 2.5rem);
+								left: calc(${ progress }% - 3.75rem);
 								top: -2.5rem;"
 						></prpl-badge>
 					</div>
@@ -135,7 +135,7 @@ const prplUpdatePreviousMonthBadgeProgressBar = ( pointsDiff ) => {
 	if ( badgeNewPoints >= parseInt( badgeMaxPoints ) ) {
 		document
 			.querySelectorAll(
-				`.prpl-badge-row-wrapper-inner .prpl-badge prpl-badge[complete="false"][badge-id="${ badgeId }"]`
+				`.prpl-badge-row-wrapper .prpl-badge prpl-badge[complete="false"][badge-id="${ badgeId }"]`
 			)
 			?.forEach( ( badge ) => {
 				badge.setAttribute( 'complete', 'true' );
@@ -161,4 +161,37 @@ const prplUpdatePreviousMonthBadgeProgressBar = ( pointsDiff ) => {
 				?.remove();
 		}
 	}
+};
+
+/**
+ * Update the previous month badge counters.
+ *
+ * @param {number} pointsDiff The points difference.
+ *
+ * @return {void}
+ */
+// eslint-disable-next-line no-unused-vars
+const prplUpdatePreviousMonthBadgeCounters = ( pointsDiff ) => {
+	const remainingPointsEls = document.querySelectorAll(
+		`.prpl-previous-month-badge-progress-bar-wrapper .prpl-previous-month-badge-progress-bar-remaining`
+	);
+
+	if ( ! remainingPointsEls.length ) {
+		return;
+	}
+
+	remainingPointsEls.forEach( ( pointsEl ) => {
+		const totalPoints = pointsEl.getAttribute(
+			'data-remaining-total-points'
+		);
+		pointsEl.setAttribute(
+			'data-remaining-total-points',
+			totalPoints - pointsDiff
+		);
+
+		const numberEl = pointsEl.querySelector( '.number' );
+		if ( numberEl ) {
+			numberEl.textContent = totalPoints - pointsDiff;
+		}
+	} );
 };