+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 
+= 1.8.0 =
+
+Added these recommendations from Ravi:
+
+* Set date format.
+
+Enhancements:
+
+* Redesign of the admin pages.
+
+Bugs we fixed:
+* Fix redirecting to Progress Planner Dashboard after login.
+
 = 1.7.2 =
 
 * Backported some features from the Progress Planner Pro plugin.